-Refactored the header drawer in its own component
