+++ src/views/_shared/header/HeaderDrawer.tsx
@@ -1,0 +1,106 @@
+import React, { useState } from "react";
+import List from "@material-ui/core/List";
+import ListItem from "@material-ui/core/ListItem";
+import ListItemText from "@material-ui/core/ListItemText";
+import SwipeableDrawer from "@material-ui/core/SwipeableDrawer";
+import MeunuIcon from "@material-ui/icons/Menu";
+import IconButton from "@material-ui/core/IconButton";
+import { makeStyles } from "@material-ui/core/styles";
+import { Link } from "react-router-dom";
+
+//import { ThemeContext } from "../../context/contexts";
+
+//styling
+const useStyles = makeStyles(theme => ({
+   toolbarMargin: {
+      ...theme.mixins.toolbar,
+      [theme.breakpoints.down("md")]: {
+         marginBottom: "3em",
+      },
+      [theme.breakpoints.down("xs")]: {
+         marginBottom: "3em",
+      },
+   },
+   drawerIconContainer: {
+      "marginLeft": "auto",
+      "&:hover": {
+         backgroundColor: "transparent",
+      },
+   },
+   drawerIcon: {
+      height: "50px",
+      width: "50px",
+      color: theme.palette.secondary.main,
+   },
+   drawer: {
+      backgroundColor: theme.palette.primary.main,
+   },
+   drawerItem: {
+      ...theme.typography.tab,
+      minWidth: "10",
+      marginLeft: "15px",
+      color: theme.palette.secondary.main,
+      opacity: 0.5,
+   },
+   drawerItemSelected: {
+      "& .MuiListItemText-root": {
+         opacity: 1,
+      },
+   },
+}));
+
+const HeaderDrawer = (props: any) => {
+    const [openDrawer, setOpenDrawer] = useState(false);
+    const classes = useStyles()
+     const { value, routes} = props 
+     const [ drawerTabValue, SetDrawerTabValue ] = useState(value)
+     //this checks if device is iOS to make the drawer swipable
+     const iOS = process.browser && /iPad|iPhone|iPod/.test(navigator.userAgent);
+
+    return (
+<>
+<SwipeableDrawer
+   classes={{ paper: classes.drawer }}
+   disableBackdropTransition={!iOS}
+   disableDiscovery={iOS}
+   open={openDrawer}
+   onClose={() => setOpenDrawer(false)}
+   onOpen={() => setOpenDrawer(true)}
+>
+   <div className={classes.toolbarMargin} />
+
+   <List disablePadding>
+      {routes.map((route:any) => (
+         <ListItem
+            key={`${route.activeIndex}`}
+            classes={{ selected: classes.drawerItemSelected }}
+            divider
+            button
+            component={Link}
+            to={route.link}
+            selected={drawerTabValue === route.activeIndex}
+            onClick={() => {
+               setOpenDrawer(false);
+               SetDrawerTabValue(route.activeIndex);
+            }}
+         >
+            <ListItemText className={classes.drawerItem}>
+               {route.name}
+            </ListItemText>
+         </ListItem>
+      ))}
+   </List>
+</SwipeableDrawer>
+<IconButton
+   className={classes.drawerIconContainer}
+   onClick={() => setOpenDrawer(!openDrawer)}
+   disableRipple
+>
+   <MeunuIcon className={classes.drawerIcon} />
+</IconButton>
+</>
+);
+
+ };
+ 
+ export default HeaderDrawer;
